requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
boto==2.49.0
boto3==1.34.10
botocore==1.34.10
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.0.4
dataclasses-json==0.6.3
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.13.3
opencv-python-headless==4.5.5.64
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
promptlayer==0.4.0
protobuf==3.19.6
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.16.2
s3transfer==0.10.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.24
streamlit==1.8.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
